New scripts for statistical analysis
The experiments can now be run with "./runExperiments.sh", and then the
output should be introduced in "stats/output_data/". Then in stats the
scripts extract_data.py and extract_stats.py should be used to get all
the stats. The directory "stats/npyFiles" should exist for these scripts to
work.